// src/model/ha/climate/service.enum.ts
export enum ClimateService {
  setAuxHeat = 'set_aux_heat',
  setFanMode = 'set_fan_mode',
  setHumidity = 'set_humidity',
  setHvacMode = 'set_hvac_mode',
  setPresetMode = 'set_preset_mode',
  setSwingMode = 'set_swing_mode',
  setTemperature = 'set_temperature',
  turnOff = 'turn_off',
  turnOn = 'turn_on',
}